+++ server/src/types/mssql-msnodesqlv8.d.ts
@@ -1,0 +1,4 @@
+declare module "mssql/msnodesqlv8" {
+  import mssql = require("mssql");
+  export = mssql;
+}